% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDER.R
\name{calculateDER}
\alias{calculateDER}
\title{Calculates the Dose Enhancement Ratio}
\usage{
calculateDER(data, control, treatment, S, method = "ml", PEmethod = "fit")
}
\arguments{
\item{data}{A data frame containing at least the following five columns: "cline", "Exp", "dose", "ncells", "ncolonies".}

\item{control}{Name of the cell-line/group that will act as control (numerator in the ratio).}

\item{treatment}{Name of the cell-line/group that will act as treatment (denominator in the ratio).}

\item{S}{The survival fraction value for which you want to calculate your DER.}

\item{method}{Method used for the fit. It's \code{"ml"} (maximum likelihood) by default. Can be \code{"ls"} (least squares) or \code{"franken"} (weighted least squares as described by Franken eta al.(2006)).}

\item{PEmethod}{Controls the value of the plating efficiencies. \code{"fit"} calculates fitted plating efficiencies as model parameters, \code{"fix"} uses fixed ones calculated from the observed zero dose data or from a column named \code{pe} in \code{data}.}
}
\value{
The Dose Enhancement Ratio of Control:Treatment.
}
\description{
It helps you calculate the Dose Enhancement Ratio without going through the data wrangling steps.
}
\examples{
calculateDER(datatab, HN31, FADU, 0.25)
calculateDER(datatab, HN31, FADU, 0.25, method = "ls", PEmethod = "fix")
}
