% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASP8_data.R
\docType{data}
\name{CASP8_data}
\alias{CASP8_data}
\title{Title}
\format{
A data frame with 144 rows and 10 variables:
\describe {
\item{cline}{}
\item{Exp}{}
\item{dose}{}
\item{ncells}{}
\item{ncolonies}{}
}
}
\source{
\url{}
}
\usage{
CASP8_data
}
\description{
Description
}
\keyword{datasets}
