% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCurves.R
\name{compareCurves}
\alias{compareCurves}
\title{Compares two cell survival curves using ANOVA.}
\usage{
compareCurves(data, cline1, cline2, method = "ml", PEmethod = "fit")
}
\arguments{
\item{data}{A data frame containing at least the following five columns:
"cline", "Exp", "dose", "ncells", "ncolonies".}

\item{cline1}{Name of the cell-line/group that you want to compare.}

\item{cline2}{Name of another cell-line/group that you want to compare.}

\item{method}{Method used for the fit. It's \code{"ml"} (maximum likelihood)
by default. Can be \code{"ls"} (least squares) or \code{"franken"}
(weighted least squares as described by Franken eta al.(2006)).}

\item{PEmethod}{Controls the value of the plating efficiencies. \code{"fit"}
calculates fitted plating efficiencies as model parameters, \code{"fix"}
uses fixed ones calculated from the observed zero dose data.}
}
\value{
An object of class \code{cellsurvLQdiff}, as returned by
  \code{\link[CFAssay:cellsurvLQdiff]{cellsurvLQdiff}}.
}
\description{
It helps compare two cell survival curves using the
\code{\link[CFAssay:cellsurvLQdiff]{cellsurvLQdiff}} function of the CFAssay
package, without having the user go through the data wrangling steps
necessary for that. Visit
\url{https://bioconductor.org/packages/release/bioc/html/CFAssay.html} for
more details about the method.
}
\examples{
datatab <- CASP8_data
compareCurves(datatab, "shCASP8-N", "shCASP8-B+Z+N")
compareCurves(datatab, "shCASP8-N", "shCASP8-B+Z+N", method = "ls", PEmethod = "fix")
}
