% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotCSCurve.R
\name{ggplotCSCurve}
\alias{ggplotCSCurve}
\title{Plots Cell Survival Curves using ggplot()}
\usage{
ggplotCSCurve(
  data,
  ...,
  method = "ml",
  PEmethod = "fit",
  colors = NULL,
  title = NULL,
  xlab = "Doses (Gy)",
  ylab = "Survival (1 = 100\%)",
  legend_title = "Cell types",
  point_shape = 16,
  point_size = 2,
  segment_type = 1,
  segment_width = 1,
  curve_width = 1,
  curve_type = 1,
  theme = ggplot2::theme_bw(),
  ylim = NULL,
  ybreaks = ggplot2::waiver()
)
}
\arguments{
\item{data}{A data frame containing at least the following five columns: "cline", "Exp", "dose", "ncells", "ncolonies".}

\item{...}{The name of the cell type(s)/group(s). If entering multiple cell types, separate each by a comma.}

\item{method}{Method used for the fit. It's \code{"ml"} (maximum likelihood) by default. Can be \code{"ls"} (least squares) or \code{"franken"} (weighted least squares as described by Franken eta al.(2006)).}

\item{PEmethod}{Controls the value of the plating efficiencies. \code{"fit"} calculates fitted plating efficiencies as model parameters, \code{"fix"} uses fixed ones calculated from the observed zero dose data or from a column named \code{pe} in \code{data}.}

\item{colors}{A vector of strings denoting the colors of the curves. Size of the vector should be equal to the number of cell types entered.}

\item{title}{A string denoting the title of the plot.}

\item{xlab}{A string denoting the label of x-axis.}

\item{ylab}{A string denoting the label of y-axis.}

\item{legend_title}{A string denoting the title of the legend.}

\item{point_shape}{An integer denoting the shape of the points in the plot.}

\item{point_size}{An integer denoting the size of the points in the plot.}

\item{segment_type}{An integer denoting the type of the straight line segments in the plot.}

\item{segment_width}{An integer denoting the width of straight line segments in the plot.}

\item{curve_width}{An integer denoting the width of curved line segments in the plot.}

\item{curve_type}{An integer denoting the type of curved line segments in the plot.}

\item{theme}{The name of the theme to be used in the plot}

\item{ylim}{A vector denoting the limits of the y-axis}

\item{ybreaks}{A vector denoting the y-axis breaks/ticks}
}
\description{
This helps to plot the cell survival curves using ggplot().
One function plots the curves, both for one cell type , or multiple cell types.
}
\examples{
# Single curve
ggplotCSCurve(datatab, "shCASP8-NT")
# This plot replicates the plot created by the CFAssay function plot.cellsurvLQfit
ggplotCSCurve(datatab, "control-B", colors = "black", xlab = "X-axis", ylab = "Y-axis", title = "Single Plot", point_shape = 1, point_size = 2, segment_width = 1, segment_type = 1, curve_width = 1, curve_type = 1, legend_title = "Legend", ylim = c(0.008, 1.05), ybreaks = c(0.01, 0.05, 0.20, 0.50, 1), theme = theme_classic())
# Multiple Curves
ggplotCSCurve(datatab, "shCASP8-N", "shCASP8-B+Z+N", "control-B", "control-NT")
ggPlotforCFAssay(datatab, "shCASP8-N", "shCASP8-B+Z+N", "control-B", "control-NT", colors = c("darkgreen", "steelblue", "orange", "magenta"), xlab = "X-axis", ylab = "Y-axis", title = "Multiple Curves", point_shape = 15, point_size = 1, segment_width = 0.5, segment_type = 1, curve_width = 1, curve_type = 1, legend_title = "Legend", theme = theme_dark())
}
