% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASP8_data.R
\docType{data}
\name{CASP8_data}
\alias{CASP8_data}
\title{CASP8 data set}
\format{
A data frame with 144 rows and 5 columns:
\describe{
    \item{cline}{}
    \item{Exp}{}
    \item{dose}{}
    \item{ncells}{}
    \item{ncolonies}{}
}
}
\usage{
CASP8_data
}
\description{
This is an example data set provided with the R package
}
\examples{
datatab <- CASP8_data

}
\keyword{datasets}
