% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmodelFit.R
\name{lqmodelFit}
\alias{lqmodelFit}
\title{Fits the Linear Quadratic Model}
\usage{
lqmodelFit(data, ctype, method = "ml", PEmethod = "fit")
}
\arguments{
\item{data}{A data frame containing at least the following five columns with
these exact names: "cline", "Exp", "dose", "ncells", "ncolonies".}

\item{ctype}{Name of the cell-line/group for which the model is to be fit.}

\item{method}{Method used for the fit. It's \code{"ml"} (maximum likelihood)
by default. Also accepts \code{"ls"} (least squares) or \code{"franken"}
(weighted least squares as described by Franken eta al.(2006)).}

\item{PEmethod}{Controls the value of the plating efficiencies. \code{"fit"}
calculates fitted plating efficiencies as model parameters, \code{"fix"}
uses fixed ones calculated from the observed zero dose data.}
}
\value{
An object of class \code{cellsurvLQfit}, as returned by
  \code{\link[CFAssay:cellsurvLQfit]{cellsurvLQfit}}.
}
\description{
It helps fit the linear quadratic model using the
\code{\link[CFAssay:cellsurvLQfit]{cellsurvLQfit}} function of the CFAssay
package, but automates the data wrangling steps, thus making it more
convenient to use for beginner R users. Visit
\url{https://bioconductor.org/packages/release/bioc/html/CFAssay.html} for
more details about the method of the fit.
}
\examples{
datatab <- CASP8_data
lqmodelFit(datatab, "control-B")
lqmodelFit(datatab, "control-B", method = "ls", PEmethod = "fix")
}
